+++ .storybook/utils/partial-action.js
@@ -1,9 +1,0 @@
-import { action } from "@storybook/addon-actions";
-
-function partialAction(actionName) {
-  return (eventObj, ...args) => {
-    action(actionName)({ ...eventObj, view: undefined }, ...args);
-  };
-}
-
-export default partialAction;