+++ src/components/pager/pager.style.tsx
@@ -81,7 +81,7 @@ const StyledPagerNavigation = styled.div`
 
   && ${StyledInputPresentation} {
     padding: 0;
-    margin: 8px 4px 0 4px;
+    margin: 4px;
     height: 26px;
     line-height: 26px;
     min-height: 24px;
@@ -98,6 +98,7 @@ const StyledPagerNavInner = styled.div`
   display: flex;
   align-items: center;
   padding: 0 12px;
+  margin: 4px 0;
 
   && ${StyledFormField} {
     margin-bottom: 0;
@@ -113,6 +114,7 @@ const StyledPagerLink = styled(Link)`
 const StyledPagerNoSelect = styled.div`
   user-select: none;
   white-space: nowrap;
+  font-weight: normal;
 `;
 
 const StyledPagerSummary = styled.div`